Adicionada a manipulação dos dados nao volateis via VEEPROM

+++ src/config/interpreter_jog_if.c
@@ -21,7 +21,7 @@ static void iif_zdown_jog(void);
 static void iif_zup_jog(void);
 static void iif_released_jog(void);
 
-extern uint32_t *velocidadeJog;
+extern float *velocidadeJog;
 char text[40];
 
 const char jog_stopflush[]= "\
@@ -49,30 +49,30 @@ void iif_esc_jog(void)
 }
 
 void iif_down_jog(void) {
-	sprintf(text, "G21 G91\nG01 Y-390.0 F%d\nm30",*velocidadeJog);
+	sprintf(text, "G21 G91\nG01 Y-390.0 F%.0f\nm30",*velocidadeJog);
 	xio_open(cs.primary_src,text,0);
 }
 void iif_up_jog(void) {
-	sprintf(text, "G21 G91\nG01 Y390.0 F%d\nm30",*velocidadeJog);
+	sprintf(text, "G21 G91\nG01 Y390.0 F%.0f\nm30",*velocidadeJog);
 	xio_open(cs.primary_src,text,0);
 }
 void iif_left_jog(void){
-	sprintf(text, "G21 G91\nG01 X-390.0 F%d\nm30",*velocidadeJog);
+	sprintf(text, "G21 G91\nG01 X-390.0 F%.0f\nm30",*velocidadeJog);
 	xio_open(cs.primary_src,text,0);
 }
 
 void iif_right_jog(void) {
-	sprintf(text, "G21 G91\nG01 X390.0 F%d\nm30",*velocidadeJog);
+	sprintf(text, "G21 G91\nG01 X390.0 F%.0f\nm30",*velocidadeJog);
 	xio_open(cs.primary_src,text,0);
 }
 
 void iif_zdown_jog(void){
-	sprintf(text, "G21 G91\nG01 Z-390.0 F%d\nm30",*velocidadeJog);
+	sprintf(text, "G21 G91\nG01 Z-390.0 F%.0f\nm30",*velocidadeJog);
 	xio_open(cs.primary_src,text,0);
 }
 
 void iif_zup_jog(void) {
-	sprintf(text, "G21 G91\nG01 Z390.0 F%d\nm30",*velocidadeJog);
+	sprintf(text, "G21 G91\nG01 Z390.0 F%.0f\nm30",*velocidadeJog);
 	xio_open(cs.primary_src,text,0);
 }
